Actually fix the failing mac agent tests.
Retry the killPid command a couple of times in case the fletch-vm is slow at getting stopped.
Ideally we should listen on the stdout of the fletch-vm and use that to determine the various states it is in, but that will be in a separate change.

R=sgjesse@google.com, ahe@google.com, karlklose@google.com
BUG=

Review URL: https://codereview.chromium.org/1427083002 .