Go back to using run.dart instead of 'fletch export' to create snapshots of the fletch agent for the SDK.
'fletch export' in the form it is used here has some side-effects
that are not desirable (creating files in home directory) and
'fletch quit' is not guaranteed currently to actually get rid
of the persistent process. It is safer to stick with run.dart
for now.

Long term we shouldn't have to use something like run.dart
for this. Our users will be using something like 'fletch export'
for scripts and we should be able to as well.

Revert "Use fletch export instead of run.dart for sdk creation."
This reverts commit 4a8031c3ebdba00b61efd9541e25e85100d9b808.

Revert "Use try-finally to make sure to run quit."
This reverts commit 863272e17cb5be4197929f10f1029f33938f543a.

R=ahe@google.com, sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1428933002 .