Do not iterate over session program roots when writing/reading snapshots
Snapshots are written via an attached session sending a
kWriteSnapshot command. Writing the snapshot involves traversing program roots,
but we should not be including the roots the session has, since it won't be
there when reading the snapshot.

Currently we write more objects into the snapshot than we will read again - due
to additional functions the session seems to be holding on to.

R=ager@google.com, ajohnsen@google.com

Review URL: https://codereview.chromium.org/1408003011 .