Add a tool to download the Fletch Raspbian image and flash it to an SD Card.
This includes a shell script that will run the tool from the Fletch SDK.

This tool will perform the following steps:

1. Determine the device for the SD Card (by asking the user to remove and insert the card)
2. Ask for hostname and IP address
3. Download the modified Raspbian image ZIP file (with progress indication)
4. Unzip the ZIP file (no progress indication there)
5. Copy the image to the SD Card  (with progress indication)
6. Write hostname and IP address to the boot partition
7. Tell the user all went well :-)

Along the way mounting, sync and un-mounting will happen behind the scenes.

I have added extensive logging, so that when this fails for somebody we can get as much information as possible.

This still have rough edges, and no testing except for running it manually. The cleanup temp dirs and mounted devices in case of error is still limited.

All the flags that can be passed are mainly for testing, maybe except --sd-card where the SD Card can be specified if the detection mechanism does not work.

Right now the modified Raspbian image is hardcoded to one I have uploaded to GCS.

BUG= https://github.com/dart-lang/fletch/issues/264
R=karlklose@google.com, ricow@google.com, wibling@google.com

Review URL: https://codereview.chromium.org/1427313004 .