Don't hang on some of the interactive debugger commands.
This changes adds a new state to the fletch session indicating whether
it is safe to ask for a stack trace. This helps us avoid hanging on
commands such as 'disasm', 'p'/'p <var>', 'bt', and 'l'.

Also did a little bit of refactoring to have only one place where we check the state of the session/program.
R=ager@google.com, ahe@google.com
BUG= https://github.com/dart-lang/fletch/issues/218

Review URL: https://codereview.chromium.org/1407063010 .