Add specialized version of the dart sdk that we can use to build documentation.
This is a temporary measure, Karl is working on generalizing the changes (and
getting it working with all fletch libraries).

This basically adds the fletch libraries to the libraries.dart file in
the sdk, so that dartdoc will be able to generate documentation. There
are still a few issues with the documentation, but for now they are
good enough.

This will not be pulled down into the normal devs checkout, but I
will add description on how to use this from local checkouts as well.

The changes to the sdk are in:
https://github.com/dart-lang/sdk/tree/_fletch_libraries

Copy of the README.google file in the bundle:
URL: https://github.com/dart-lang/sdk/tree/_fletch_libraries
Version: 1c14566853da8ffdd4a41ee5d4676ff437370c2d
License: BSD
License File: LICENSE

Description:
This bundle includes a sligtly modified dart-sdk. The changes can be seen in
the github repo specified as URL above. It allows us to create documentation
for the fletch packages. The actual libs from fletch needs to be copied in
under lib/mobile for this to work.

Local Modifications:
None

R=sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1412653007 .