Implement optimistic deallocation of the latest allocated object for mark-sweep.
This is used for the ffi natives. A pointer doesn't necessarily fit in
a smi and therefore the pattern for the FFI natives that return a
pointer is:

LargeInteger* i = NewLargeInteger();
if (i == alloc_failure) return i;
// compute result (can have side-effects)
if (Smi::IsValid(result)) {
  TryDealloc();
  return result;
}
i.set_value(result);
return i;

Most of the time we don't need the large integer, but because of the
retry on allocation failure we have to allocate it up front to not
repeat side-effects when retrying. We added TryDealloc to make this
particular pattern cheap.

R=erikcorry@google.com

Review URL: https://codereview.chromium.org/1437923002 .