Don't abort in the spawned child unless we fail to report the error.
This should get rid of the core dumps we are seing on the bots

Fix native_process spawn to do an immediate abort.

ScheduleAbort requires a follow-on exit which I had missed. Either the previous FATAL did an ImmediateAbort if kind == ASSERT so we will just call that here as well.

R=kustermann@google.com, sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1446213002 .