Move writing to stdout out session methods for the 'p' debugger command.
I am in the process of refactoring the session.dart file to not have it
do any printing to stdout, but rather return the result and leave the
printing to the caller. This is the first step.

Also changed the so/sob to n/nb as having a single letter for stepping a
line is much nicer that two and 'n' is used in gdb.

R=ahe@google.com, ager@google.com, sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1461593002 .