Add utility function to allow easily looking up a variable
This was already possible, but this would require doing something like:
  ForeignLibrary fl = new ForeignLibrary.fromName(libPath);
  ForeignFunction count = fl.lookup('count');
  ForeignMemory countMem = new ForeignMemory.fromAddress(countMem.address);
  var val = countMem.getInt32();

(or pass in the foreignfunction as a parameter)

Intuitively, the above does not play nicely, why is it a foreign function?

This will make us instead do:
  ForeignLibrary fl = new ForeignLibrary.fromName(libPath);
  ForeignPointer countPointer = fl.lookupVariable('count');
  ForeignMemory countMem = new ForeignMemory.fromAddress(count.address, 4);
  var val = countMem.getInt32();

Not much of a difference, but for sanity I think it is much better.

Note that, in many cases, we will not even do the last two lines
above, we will simply use the varible as a parameter to another
foreign call (and in many cases it will be a struct, not a simple
value)

BUG=
R=sgjesse@google.com

Review URL: https://codereview.chromium.org/1464393002 .