Make lookup of our default .fletch-settings template a bit more robust.
This change looks up in the parent directory chain of the fletch
executable to find a .fletch-settings file in addition to the current
working directory. I changed it to also look for
$sessionName.fletch-settings file starting from current working
directory and up its parent chain.
The latter means we can now add a foo.fletch-settings in a directory and
have that be used as the project configuration.

R=sgjesse@google.com, ahe@google.com, karlklose@google.com
BUG=

Review URL: https://codereview.chromium.org/1474693004 .