Create debugger structures in the VM whenever we need them.
When I added debugger interruption I didn't take care of the fact that
we can now stop without having set a breakpoint. Therefore, the VM
debugger datastructures are not setup and an attempt to perform a
stepping operation will crash the VM.

After this patch we setup the debugger datastructures whenever we
need them.

This is a little hard to reproduce in our test framework because
it relies on the Ctrl-\ interruption mechanism. As a next step
we need to get this tested.

R=zerny@google.com

Review URL: https://codereview.chromium.org/1476863004 .