Propagate signals first and monitor exit signals later
After this CL we have a concept of a spawning tree. Each process keeps track of it's parent. If a process dies, it sends signals to all linked processes ... But if one process monitors another process it will only get a message once the monitored process and **it's process tree** are gone.

For example:

[main]
  |-monitor-> [eventbus]
  |-monitor-> [ui]
               |<-link-> [renderer]
               |<-link-> [display]

If [ui] dies, it will signal [renderer] and [display] to die and once all of them are dead, [main] will get notified an can restart [ui]. This *ensures* a whole sub-tree is dead before restarting it.

Another nice benefit is that the exitcode will now be always defined by how [main] exited, since it is the last process to die (we have always links downwards, so if [main] dies, it's whole subtree is cleaned up first)

R=kasperl@google.com

Review URL: https://codereview.chromium.org/1453913004 .