Refactor most of session.dart to not print to stdout.
This change moves most of the printing to stdout out of session.dart.
Instead we now return information to the caller about what happened.

This does make input_handler and debug_verbs able to handle output
differently, but also means there is some duplication.
To keep it somewhat down I added a session.processStopResponseToString
method that can be used to convert a response to a default response
string.

This change also changes the standard output for most of the debugger
commands. Instead of printing a stack trace we now print the current
location in the source code after stepping, hitting a breakpoint etc.
This is more inline with gdb's output. The user will specifically have
to ask for a stack trace using the 'bt' command.

I have not yet updated the expected test data, since I wanted people's
feedback first.

R=lukechurch@google.com, zerny@google.com, ahe@google.com, sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1475053004 .