Compile fletch with -fno-strict-aliasing.
The type of VM code that we are writing is inherently incompatible with
strict aliasing optimizations. We are operating on raw memory that is
aliased with multiple different pointer types and we need to do so.

Using bit_cast<T*> only silences warnings but does not prevent the
strict aliasing optimizations. Therefore, I removed our usages
of bit_cast on pointer types.

R=erikcorry@google.com, ajohnsen@google.com

Review URL: https://codereview.chromium.org/1496893002 .