Store FFI foreign pointers as two Smis to avoid external heap number.
The heap number is problematic for two reasons.  The first is that
we end up looking at two different foreign objects during finalization.
This is tough on the GC because it means the objects have to die at the
same time.  The second is that it causes double allocations, which are
problematic with the retry-GC semantics.
R=kustermann@google.com
BUG=

Review URL: https://codereview.chromium.org/1492943004 .