A bunch of renaming and commenting on the persistent process code.
The naming is based on four entities.

1. The Fletch C++ client (called client in the code)
2. The main isolate of the persistent process (called hub (main
isolate))
3. Worker isolates in the persistent process (called worker isolate)
4. The Fletch VM.

There are now ClientCommands that are sent between the Fletch C++ client,
the hub (main isolate), and worker isolates. There are also VmCommands
that are sent between the worker isolate and the Fletch Vm.

I will make a subsequent change that renames the files and moves some of them into a new lib/src/worker directory.

R=karlklose@google.com, sgjesse@google.com, ahe@google.com
BUG=

Review URL: https://codereview.chromium.org/1512893002 .