Add verification that natives only perform one allocation.
The only place where we did more than one allocation was in connection
with process death. I split that into two native calls for now. That
is not great as it slows down message handling, but I think it is
the right thing to do for now.

Next step is to encode the requested amount in the allocation failure
objects and fail with out of memory if we cannot get that amount of
memory after a GC on allocation failure.

R=kustermann@google.com, kasperl@google.com

Review URL: https://codereview.chromium.org/1523233002 .