Delete tests/fletch/.packages
It was not complete, did not parse with the analyzer, and the tests
are always run with the .packages in the root of the repo.

This allows the analyzer to pick up the .packages in the root of the
repo when editing the tests in intellij/atom.

BUG=
R=sgjesse@google.com

Review URL: https://codereview.chromium.org/1522203003 .