MultiProgram support Part 1: Let the GC thread notify the scheduler once it is done with a program
The scheduler needs to know when a program is done running. As long as the GC
thread has a reference to the program it cannot be deleted. (This didn't matter
before, because the Scheduler::Run() method would only return once the GC thread
was Join()ed and the program deletion happened after Scheduler::Run(). Multiprogram
support will change this fact.)

R=herhut@google.com

Review URL: https://codereview.chromium.org/1533083002 .