MultiProgram support Part 2: Make scheduler a background service
This CL makes the scheduler a background service. This is the base work for
allowing any client to talk to the VM, schedule a new program and get notified
once that program is done running.

In order to support this the following this CL makes a number of changes:
  * Scheduler::Run() does not just halt after `process_count() == 0`. Instead it
    runs as long as the VM is not turned down.

  * The embedder gets a callback once a program has died (sucessfully or
    unsuccessfully)

  * The scheduler keeps track of users of a Program via
    Program::ProgramState::refcount_.

  * Adds a SimpleProgramRunner, which gets an array of programs (and potentially
    main processes) waits for all of them to exit and then returns.

R=herhut@google.com

Review URL: https://codereview.chromium.org/1537933002 .