Add a circular buffer to the ffi package usable from both dart and c
I have not included a c library for this in this cl (only some sanity checking functions in the c ffi_test_library.
Not completely sure how to make such a library available and where to put it, but I think it is good to review the two sides separately anyway.

BUG=
R=sgjesse@google.com

Review URL: https://codereview.chromium.org/1561823004 .