Add Dart code for computing binary size information on Golem.
This relies on Golem builders producing two files using the 'size'
command on Linux. One for a baseline system containing LK and one
for the LK system also containing Fletch. The script in this
change will compute the size that Fletch adds to the image.

R=ricow@google.com

Review URL: https://codereview.chromium.org/1573243013 .