Stop using the pattern 'ldr register, =constant' for 8 bit values.
The checked in GCC will convert any such 'ldr' instruction with a
constant that is between -256 and 255 to a 'cmp' instruction. That
seems funky.

The manual says

(http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.dui0204j/Chdcegci.html)

"""
The LDR Rd,=const pseudo-instruction can construct any 32-bit numeric
constant in a single instruction. You can use this pseudo-instruction to
generate constants that are out of range of the MOV and MVN
instructions.
"""

With this change we are only using this pseudo instruction when
out of range for 'mov' and 'mvn'.

R=erikcorry@google.com, ricow@google.com

Review URL: https://codereview.chromium.org/1603843003 .