Add page allocator and new C/C++ heap to disco_fletch
Add a page allocator with a C-malloc implementation on top. The
page allocator is mostly ported from the C implementation added
to the LK kernel project and the C-malloc implementation is
cmpctmalloc also from the LK kernel project.

The cmpctmalloc code is still in C and have #ifdef/#endif to
port the code.

During startup the page allocator is initialized before C++
static initializers are run. This is done by wrapping
the __libc_init_array using the GNU linker --wrap option. This is
done this way mainly to avoid changing the startup code in
startup_stm32f746xx.s.

The newlib malloc implementation is no longer used, and is
overridden by wrapping _malloc_r, _realloc_r, _calloc_r and
_free_r again using the GNU linker --wrap option. The
implementation of _sbrk is invalidated, as it should no longer be
called - all allocation must go through the page allocator.

The page allocator is initialized to system RAM, and after
startup the external RAM is added as a separate arena where
currently only the screen frame buffer is allocated from.

R=erikcorry@google.com, ricow@google.com

Review URL: https://codereview.chromium.org/1607793003 .