Add fletch package for using tls client sockets.
This is using the mbed tls c library that we already build as part of
our build process (as a shared library).

This is currently a little limited, in that you have no control over
certificate etc, but we can iterate on this in the repo.

This has been also used sort of a case study for the ffi library, and
is as such not optimized (we could bring a lot of the code in the
constructor into the c side and make this smaller and more efficient)

BUG=
R=sgjesse@google.com

Review URL: https://codereview.chromium.org/1605373002 .