Update the SDK to include support for STM32F746G Discovery board
The script to bundle the SDK will now include the required static
libraries for buinding an image for the STM32F746G Discovery board.

Added a script to build the static image. This script will be the basis
for adding building support to the fletch command.

R=karlklose@google.com
BUG=

Review URL: https://codereview.chromium.org/1645033002 .