Add an x-build command
The x-build command will compile the code into a snapshot, and then link
it into an image which can be flashed on a STM32F746G Discovery board.

The actual linking of the image is handled by calling a shell script.

Running the x-build command should work in both a Git chackout and
an SDK.

R=ricow@google.com, ahe@google.com
BUG=

Review URL: https://codereview.chromium.org/1674533002 .