Add additional kEnqueuing state to Process.
We now transition to kEnqueuing when a process is about to be added to the ready queue. The queue itself then adds the process and changes the state, after adding, to kReady. This should avoid races on queue manipulation based on a process' state.

BUG=
R=kustermann@google.com

Review URL: https://codereview.chromium.org/1699093002 .