update _embedder.yaml to reflect repo layout
I'd like to open the repo/sdk/pkg directory in Atom
and have it correctly analyzed.

To that end, this changes the _embedder.yaml file so that it can be used
for analysis while in the repo tree structure. In addition, I modified
the bundle_sdk.py script to update the _embedder.yaml file during the
build process to reflect the Dartino SDK structure.

R=sgjesse@google.com

Review URL: https://codereview.chromium.org/1727233004 .