Improve the ARM exception handler when running under FreeRTOS
Use the correct stack pointer psp/msp when finding the exception frame.

Move the non-exception frame registers to a separate structure as they
might not be on the same stack as the exception frame.

Add some volatile locals for easier state inspection in GDB.

Disable printing the exception for now, as the printf code will most
likely fail when called from the exception hander.

R=erikcorry@google.com
BUG=

Review URL: https://codereview.chromium.org/1767643002 .