LK fixes for large-heap tests
Two things going on here.  The new-space is now limited to 16k which slows
things down a bit for very large heaps, but I think it's the right thing for
the hardware we are targeting.  This also causes more fragmentation, which
is something we need to look at:  The freelist allocation is a bit too simple-
minded right now.
R=sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/1776993005 .