Allow limited debugging of a vm running from a snapshot.
By reading the programInfo from disk, the offset of the entry function can be pushed on the vm's session stack, and the program can be
run.

The hash of a snapshot is now based on the actual bytes of the snapshot, and the hash is checked to ensure the same snapshot is being debugged, as the debug-info is created from.

Still there no support for breakpoints etc.

R=erikcorry@google.com, sgjesse@google.com

Review URL: https://codereview.chromium.org/1754203002 .