Handle mark stack overflow.
During a mark-sweep GC the work-list can overflow. Until now we just grew
it with a new malloc'ed allocation during GC. However it is unhealthy to
allocate memory during GC, the exact moment when memory pressure is at its
highest.

We maintain a bit per card (0.1% overhead) which tells us whether the card
contains objects we could not push on the stack due to mark stack overflow.
Because the bitmap is only 0.1% of the heap size, it is very fast to scan the
heap for grey (live-but-not-yet-scanned) objects that should be scanned.

Using this, we can reduce the mark stack size from unbounded to 128 words
0.5k without noticable slowdown.
R=kustermann@google.com
BUG=

Review URL: https://codereview.chromium.org/1851033002 .