Add compactor to old-space GC
This sliding compactor compacts the whole heap in one stop-the-world GC without using any spare memory, other than the GC metadata, which is now 8% of the heap size.  The memory can be spread over several disjoint chunks (which complicates the implementation quite a bit).

It uses a single pass over the objects (and some metadata accesses) and does not need virtual memory support.  In comparison, the Jonkers pointer-reversal scheme, which is also in-place and also does not need VM, has two passes and performs noticeably worse in my tests.

I investigated the performance of partial compaction where eg 50% of the heap is moved and the other 50% stay put. The performance benefit was not very convincing, so I decided to keep it simple, either compacting or mark-sweeping-with-free-list. The current heuristic is to alternate between these two forms, but I fully expect to do something smarter at some point.

As a rule of thumb if the free-list based GC can mark-sweep the whole heap at a speed of x kb/ms then the compactor can compact the whole heap at 0.7x kb/ms. This CL also speeds up the free-list based GCs by about 17% using prefetches, inlining and other tweaks. Measurements were done on mini-Splay on embedded hardware using the live data size as the metric.
R=kustermann@google.com
BUG=

Review URL: https://codereview.chromium.org/1945793003 .