Optionally include foreign libraries when building embedded deployment
This adds a simple way of including additional libraries (.a or .o
files) when running 'dartino build' and 'dartino flash'. All files in
the ffi subdirectory relative to the .dart file being compiled are added
to the link step.

Added a sample of how to compile a native library and an additional
static ffi symbol file.

Also added an intrinsic set of static ffi functions (currently only
strlen) because the ffi package use C function from the main library.

This needs to be further refined.

Closes #482

BUG= https://github.com/dartino/sdk/issues/482
R=karlklose@google.com

Review URL: https://codereview.chromium.org/1971243002 .