Remove bogus asserts
The idea was to assert that objects are grey before we scan them and turn
them black.  It turns out this is not always the case.  When there is a
mark stack overflow we sometimes find grey objects by heap iteration
that are already on the marking stack.  We scan and blacken them during
the heap iteration, and when they are later popped, they get harmlessly
scanned again.
R=sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/2017603002 .