More compact snapshots.
Uses a variable-length encoding for the w and x parts of the snapshot
byte code. Does more work at snapshot creation time to find a more
optimal choice of bytecodes to encode the snapshot.

Also arranges for the null, true, and false objects to be in fixed positions, relative to each other. This frees up 2 registers in the ARM interpreter, which we use for the program and the dispatch table.

R=sgjesse@google.com
BUG=

Review URL: https://codereview.chromium.org/2067483002 .