Add I2C driver for F4 Nucleo board.
The F4 MCU has completely different I2C hardware from the F7, so this is
a new implementation, based on the F4 HAL.

Also contains a tiny cleanup of the F7 I2C driver. The read/write
direction is set in the flags, and the LSB of the address (SADD0) is
ignored by the hardware.

BUG=
R=karlklose@google.com

Review URL: https://codereview.chromium.org/2186863004 .