Add file system driver model
This adds the FileSystem class in which several file systems can be
mounted. Each file system is implemented as a an instance of
FileSystemDriver.

A read only file system is implemented in the ReadOnlyFileSystemDriver.

This has tests, but is not yet hooked up to anything.

R=erikcorry@google.com, jakobr@google.com
BUG=

Review URL: https://codereview.chromium.org/2214283004 .