Add flags -Xassume-no-nsm and -Xwrap-smis to llvm-codegen
These flags cheat in order to evaluate the potential of the LLVM backend. The
first assumes that strong mode has eliminated the possibility of no-such-method
on all calls. The second is allows Smis to wrap without promoting them to big
numbers (but even with this flag, big numbers still exist, complicating the
dispatch of arithmetic operators).
TBR=dmitryolsh@google.com
BUG=

Review URL: https://codereview.chromium.org/2469963006 .