Compile statically monomorphic calls as a check and a call
Also:
Compile subtree-of-classes dispatches to a classid range check
Improve array load code

With this change, we throw an exception (42) on no-such-method. This is not
quite the correct behaviour, but we judge it to be just as performant as adding
actual NSM support, so that we can get performance numbers without having to do
all the work.
BUG=

Review URL: https://codereview.chromium.org/2533113004 .